+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 0.0.2
 * Add a customisation screen to select the lobby/skyblock in Create New World, More Options (warjort)
 * Add an example lobby config (warjort)
+* Save all major config in the persistant data, only skyblocks are re-read after world creation (warjort)
 
 ### 0.0.1
 * Initial release